refactor: Move QQGroupEntity into QQBotEntity
